Extract Apps action group actions into standalone class
