+++ src/codemagic/tools/_app_store_connect/action_groups/__init__.py
@@ -1,1 +1,2 @@
 from .app_store_action_group import AppStoreActionGroup
+from .apps_action_group import AppsActionGroup
