+++ src/codemagic/tools/app_store_connect.py
@@ -18,9 +18,6 @@ from codemagic.apple import AppStoreConnectApiError
 from codemagic.apple.app_store_connect import AppStoreConnectApiClient
 from codemagic.apple.app_store_connect import IssuerId
 from codemagic.apple.app_store_connect import KeyIdentifier
-from codemagic.apple.resources import App
-from codemagic.apple.resources import AppStoreState
-from codemagic.apple.resources import AppStoreVersion
 from codemagic.apple.resources import AppStoreVersionSubmission
 from codemagic.apple.resources import BetaAppReviewSubmission
 from codemagic.apple.resources import BetaReviewState
@@ -46,6 +43,7 @@ from codemagic.models import PrivateKey
 from codemagic.models import ProvisioningProfile
 
 from ._app_store_connect.action_group import AppStoreConnectActionGroup
+from ._app_store_connect.action_groups import AppsActionGroup
 from ._app_store_connect.action_groups import AppStoreActionGroup
 from ._app_store_connect.arguments import AppArgument
 from ._app_store_connect.arguments import AppStoreConnectArgument
@@ -76,9 +74,10 @@ def _get_certificate_key(
 
 @cli.common_arguments(*AppStoreConnectArgument)
 class AppStoreConnect(cli.CliApp,
-                      PathFinderMixin,
+                      AppsActionGroup,
                       AppStoreActionGroup,
-                      ResourceManagerMixin):
+                      ResourceManagerMixin,
+                      PathFinderMixin):
     """
     Utility to download code signing certificates and provisioning profiles
     from Apple Developer Portal using App Store Connect API to perform iOS code signing
@@ -125,56 +124,6 @@ class AppStoreConnect(cli.CliApp,
             **cls._parent_class_kwargs(cli_args),
         )
 
-    @cli.action('list',
-                BundleIdArgument.BUNDLE_ID_IDENTIFIER_OPTIONAL,
-                AppArgument.APPLICATION_ID_RESOURCE_ID_OPTIONAL,
-                AppArgument.APPLICATION_NAME,
-                AppArgument.APPLICATION_SKU,
-                AppStoreVersionArgument.VERSION_STRING,
-                AppStoreVersionArgument.PLATFORM,
-                AppStoreVersionArgument.APP_STORE_STATE,
-                action_group=AppStoreConnectActionGroup.APPS)
-    def list_apps(self,
-                  bundle_id_identifier: Optional[str] = None,
-                  application_id: Optional[ResourceId] = None,
-                  application_name: Optional[str] = None,
-                  application_sku: Optional[str] = None,
-                  version_string: Optional[str] = None,
-                  platform: Optional[Platform] = None,
-                  app_store_state: Optional[AppStoreState] = None,
-                  should_print: bool = True) -> List[App]:
-        """
-        Find and list apps added in App Store Connect
-        """
-
-        apps_filter = self.api_client.apps.Filter(
-            bundle_id=bundle_id_identifier,
-            id=application_id,
-            name=application_name,
-            sku=application_sku,
-            app_store_versions=version_string,
-            app_store_versions_platform=platform,
-            app_store_versions_app_store_state=app_store_state,
-        )
-        return self._list_resources(apps_filter, self.api_client.apps, should_print)
-
-    @cli.action('get', AppArgument.APPLICATION_ID_RESOURCE_ID, action_group=AppStoreConnectActionGroup.APPS)
-    def get_app(self, application_id: ResourceId, should_print: bool = True) -> App:
-        """
-        Get information about a specific app
-        """
-
-        return self._get_resource(application_id, self.api_client.apps, should_print)
-
-    @cli.action('builds', AppArgument.APPLICATION_ID_RESOURCE_ID, action_group=AppStoreConnectActionGroup.APPS)
-    def list_app_builds(self, application_id: ResourceId, should_print: bool = True) -> List[Build]:
-        """
-        Get a list of builds associated with a specific app
-        """
-
-        return self._list_related_resources(
-            application_id, App, Build, self.api_client.apps.list_builds, None, should_print)
-
     @cli.action('pre-release-version',
                 BuildArgument.BUILD_ID_RESOURCE_ID,
                 action_group=AppStoreConnectActionGroup.BUILDS)
@@ -185,52 +134,6 @@ class AppStoreConnect(cli.CliApp,
 
         return self._get_related_resource(
             build_id, Build, PreReleaseVersion, self.api_client.builds.read_pre_release_version, should_print)
-
-    @cli.action('pre-release-versions',
-                AppArgument.APPLICATION_ID_RESOURCE_ID,
-                action_group=AppStoreConnectActionGroup.APPS)
-    def list_app_pre_release_versions(
-            self, application_id: ResourceId, should_print: bool = True) -> List[PreReleaseVersion]:
-        """
-        Get a list of prerelease versions associated with a specific app
-        """
-
-        return self._list_related_resources(
-            application_id, App, PreReleaseVersion, self.api_client.apps.list_pre_release_versions, None, should_print)
-
-    @cli.action('app-store-versions',
-                AppArgument.APPLICATION_ID_RESOURCE_ID,
-                AppStoreVersionArgument.APP_STORE_VERSION_ID_OPTIONAL,
-                AppStoreVersionArgument.VERSION_STRING,
-                AppStoreVersionArgument.PLATFORM,
-                AppStoreVersionArgument.APP_STORE_STATE,
-                action_group=AppStoreConnectActionGroup.APPS)
-    def list_app_store_versions_for_app(
-            self,
-            application_id: ResourceId,
-            app_store_version_id: Optional[ResourceId] = None,
-            version_string: Optional[str] = None,
-            platform: Optional[Platform] = None,
-            app_store_state: Optional[AppStoreState] = None,
-            should_print: bool = True) -> List[PreReleaseVersion]:
-        """
-        Get a list of prerelease versions associated with a specific app
-        """
-
-        app_store_versions_filter = self.api_client.app_store_versions.Filter(
-            id=app_store_version_id,
-            version_string=version_string,
-            platform=platform,
-            app_store_state=app_store_state,
-        )
-        return self._list_related_resources(
-            application_id,
-            App,
-            AppStoreVersion,
-            self.api_client.apps.list_app_store_versions,
-            app_store_versions_filter,
-            should_print,
-        )
 
     @cli.action('list-builds',
                 AppArgument.APPLICATION_ID_RESOURCE_ID_OPTIONAL,
